requirements.txt
aiogram==3.3.0
Pillow==10.1.0